improve code
