+++ apm-sniffer/apm-toolkit-activation/apm-toolkit-trace-activation/src/main/resources/skywalking-plugin.def
@@ -16,9 +16,7 @@
 
 toolkit-trace=org.apache.skywalking.apm.toolkit.activation.trace.ActiveSpanActivation
 toolkit-trace=org.apache.skywalking.apm.toolkit.activation.trace.TraceAnnotationActivation
-toolkit-trace=org.apache.skywalking.apm.toolkit.activation.trace.TraceStaticAnnotationActivation
 toolkit-trace=org.apache.skywalking.apm.toolkit.activation.trace.TraceContextActivation
 toolkit-trace=org.apache.skywalking.apm.toolkit.activation.trace.CallableOrRunnableActivation
 toolkit-tag=org.apache.skywalking.apm.toolkit.activation.trace.TagAnnotationActivation
-toolkit-tag=org.apache.skywalking.apm.toolkit.activation.trace.TagStaticAnnotationActivation
 toolkit-exception=org.apache.skywalking.apm.toolkit.activation.trace.IgnoredExceptionAnnotationActivation
